Frag transaction

+++ PerfectMe/app/src/main/java/com/singularitycoder/perfectme/RoutineStep.kt
@@ -1,7 +1,7 @@
 package com.singularitycoder.perfectme
 
 data class RoutineStep(
-    val stepNumber: Short,
+    val stepNumber: Int,
     val stepName: String,
     val stepDuration: String,
 )
